Trying to tone down deprecation warnings.
LSP4J helpfully deprecates a ton of stuff that is nevertheless a part of their
main API in the form of Eithers. So there's no way to prevent the types from
appearing in a ton of expressions even if we _wanted_ to stop using them, which
we don't. Because there are a zillion ancient LSP servers out there that will
never update to the latest crap, so we have to support every half-assed,
ill-conceived step on LSP's plodding path to Nirvana.