Machinery for using tree-sitter for code highlighting.
Right now it's wildly inefficient, reparsing and restyling the entire buffer on
every keystroke. Lol! There are many steps toward efficiency that I can take,
with increasing degrees of PITA:

1. Reparse the whole buffer with tree-sitter on change, but only update the
styles and syntaxes of the lines that changed.

2. Save the tree-sitter tree from each parse, use the tree edit mechanism to
update that tree when buffer changes come in, and then pass in the old tree to
presumably make the reparsing process more efficent. I think this is possible
with the current Java tree-sitter bindings.

3. Figure out how to wrangle JNI into wrapping a C API that takes a pointer to
a callback function which returns a pointer into unmanaged memory to enable
tree-sitter to operate on (as close as possible to) the raw data for each line
stored in MutableLine's internal char arrays, instead of having to concatenate
the entire buffer into a string and convert that to UTF-16 bytes every time we
make a single change to the buffer. This kind of two-way back and forth across
the VM boundary is difficult and annoying, but OMG turning the whole buffer
into a string on every keystroke is egregious.

Also all of this relies on a local build of java-tree-sitter because the
published build does not work on M1 Macs, yay. But if I have to hack that
project's JNI bindings, I'll be needing my own version anyway, at least until
such time as I can get the additions merged upstream and fix their build to
build a fat library for macOS. Fun? No, not fun.