Provide a place for installing "extra" stuff.
The Eclipse langclient for example needs to download a bunch of Eclipse code
and put it somewhere. So now it has a reasonable place for that to go.

Having this in PackageService is kind of random. And I should probably just get
rid of the indirection in service implementations, but anyway...